moved measurement files into their own package
