+++ app/src/main/java/com/kronos/skilletapp/model/measurement/Measurement.kt
@@ -1,9 +1,8 @@
-package com.kronos.skilletapp.model
+package com.kronos.skilletapp.model.measurement
 
 import com.github.michaelbull.result.*
 import com.kronos.skilletapp.utils.roundToEighth
 import kotlin.math.roundToInt
-import com.kronos.skilletapp.model.IngredientType.*
 import com.kronos.skilletapp.utils.haveSameTypes
 import com.kronos.skilletapp.utils.toFraction
 import kotlinx.serialization.Serializable
