+++ app/src/main/java/com/kronos/skilletapp/model/measurement/MeasurementConverter.kt
@@ -1,4 +1,4 @@
-package com.kronos.skilletapp.model
+package com.kronos.skilletapp.model.measurement
 
 interface MeasurementConverter {
   val ratio: Float
