+++ app/src/main/java/com/kronos/skilletapp/parser/IngredientVisitor.kt
@@ -1,8 +1,8 @@
 package com.kronos.skilletapp.parser
 
 import com.kronos.skilletapp.model.Ingredient
-import com.kronos.skilletapp.model.Measurement
-import com.kronos.skilletapp.model.MeasurementUnit
+import com.kronos.skilletapp.model.measurement.Measurement
+import com.kronos.skilletapp.model.measurement.MeasurementUnit
 import com.kronos.skilletapp.parser.grammar.IngredientGrammarBaseVisitor
 import com.kronos.skilletapp.parser.grammar.IngredientGrammarParser
 import com.kronos.skilletapp.utils.Fraction
